78079	Exception thrown out when edit parameter's default value in chart dialog 78125	BIRT Chart width of 100% shrinks to 0 78433	Validation of chart scale value should be enhanced 77743 (partial)	set chart html output file invisible will cause whole report invisible 76910	Checkbox Strikethrough doesn't take effect

+++ chart/org.eclipse.birt.chart.reportitem/src/org/eclipse/birt/chart/reportitem/ChartReportItemPresentationImpl.java
@@ -330,15 +330,18 @@ public final class ChartReportItemPresentationImpl extends ReportItemPresentatio
         }
 
         // DELETE THE TEMP CHART IMAGE FILE CREATED
-        if (!fChartImage.delete())
+        if ( fChartImage != null )
         {
-            DefaultLoggerImpl.instance().log(ILogger.ERROR,
-                "Could not delete temporary " + sExtension + " file created at " + fChartImage.getPath());
-        }
-        else
-        {
-            DefaultLoggerImpl.instance().log(ILogger.INFORMATION,
-                "Successfully deleted temporary " + sExtension + " file created at " + fChartImage.getPath());
+	        if (!fChartImage.delete())
+	        {
+	            DefaultLoggerImpl.instance().log(ILogger.ERROR,
+	                "Could not delete temporary " + sExtension + " file created at " + fChartImage.getPath());
+	        }
+	        else
+	        {
+	            DefaultLoggerImpl.instance().log(ILogger.INFORMATION,
+	                "Successfully deleted temporary " + sExtension + " file created at " + fChartImage.getPath());
+	        }
         }
         DefaultLoggerImpl.instance().log(ILogger.INFORMATION, "ChartReportItemPresentationImpl: finish(...) - end");
     }
